+++ MiDAS/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.0.9004 - 18/03/2019
++ introduce forwardAllelesSelection for stepwise conditional testing, adding the 
+  previous top-associated allele as covariate, until there’s no more significant 
+  alleles 
+
 ## 0.0.0.9003 - 05/02/2019
 + G groups alleles are now accepted as properly formatted
 + G groups alleles can be now reduced, with warning